+++ esp-bootloader-esp-idf/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `OtaUpdater::next_partition` (#4339)
 
 ### Changed